project/requirements.txt
celery==4.4.7
Django==3.2.4
djangorestframework
redis==3.5.3
requests==2.27.1

pytest==6.2.4
pytest-django==4.4.0
